+++ app/src/Containers/EventSearchContainer.tsx
@@ -2,15 +2,21 @@ import React, { useState } from 'react'
 import dayjs from 'dayjs'
 import styled from 'styled-components/native'
 import { ScrollView, SafeAreaView } from 'react-native'
+import { useSelector } from 'react-redux'
 import TopOfApp from '@/Components/TopOfApp'
 import underDevelopmentAlert from '@/Utils/UnderDevelopmentAlert'
 import EventSearchCalendarPicker from '@/Components/Event/EventSearchCalendarPicker'
-import EventSearchUpcomingQuickSearch, {
+import EventSearchQuickSearch, {
   EventQuickSearchChoice,
-} from '@/Components/Event/EventSearchUpcomingQuickSearch'
+} from '@/Components/Event/EventSearchQuickSearch'
+import EventSearchQuickSearchUpcoming, {
+  EventQuickSearchUpcomingChoice,
+} from '@/Components/Event/EventSearchQuickSearchUpcoming'
 import QuickSearchButton from '@/Components/Forms/QuickSearchButton'
 import FreeSearchBar from '@/Components/FreeSearchBar'
 import { Events } from '@/Services/modules/events'
+import { EventsState } from '@/Store/Events'
+import { fuzzySearchByArray } from '@/Utils/Search'
 
 const Heading = styled.Text`
   font-weight: bold;
@@ -35,10 +41,10 @@ const EventTitle = styled.Text`
 
 export interface EventSearchInterface {
   type: 'date' | 'text' // what type of search is it
-  quickSearchChoice?: EventQuickSearchChoice // quick search (if any)
-  range: 'past' | 'upcoming' // which range of events are being searched
+  range: 'all' | 'past' | 'upcoming' // which range of events are being searched
   results: Events // the events results for this search
   description?: string // some text to tell the user what the search was, e.g. the date range
+  quickSearchUpcomingChoice?: EventQuickSearchUpcomingChoice // upcoming date quick search (if any)
 }
 
 export const filterEventsByDate = (
@@ -65,28 +71,53 @@ export const filterEventsByDate = (
   })
 
 const EventSearchContainer = () => {
-  const PopularSearches = [
-    'Talk to the hiring manager',
-    'Atlassian huddle',
-    'Orientation',
-    'Showcase',
-  ]
+  const [calendarPickerWidth, setCalendarPickerWidth] = useState(0)
+
+  // Get all upcoming events from the Redux store (these are added to the store by the EventsContainer component)
+  const allUpcomingEvents = useSelector(
+    (state: { events: EventsState }) => state.events.upcoming,
+  )
+
+  // TODO: when past events are also being brought into the app, change the line below
+  // to combine allUpcomingEvents and past events e.g. const allEvents = [...allUpcomingEvents, ...allPastEvents]
+  const allEvents = [...allUpcomingEvents]
+
+  const getEventsSeriesChoices = (): EventQuickSearchChoice[] => {
+    const eventsSeriesChoices = [] as EventQuickSearchChoice[]
+
+    for (const event of allEvents) {
+      if (
+        event.series &&
+        !eventsSeriesChoices.some(choice => choice.value === event.series)
+      )
+        eventsSeriesChoices.push({
+          text: event.series,
+          value: event.series,
+        })
+    }
+
+    eventsSeriesChoices.sort()
+
+    return eventsSeriesChoices
+  }
+
+  const handleFreeTextChange = (input: React.SetStateAction<string>) => {
+    console.log(input)
+  }
+
+  const handleFreeTextSubmit = () => {
+    console.log('Submit')
+  }
+
+  // Define options for quick search buttons
+  const eventsSeriesChoices = getEventsSeriesChoices()
+
   const ProjectSearches = [
     'Sole drop-in',
     'Volunteering app',
     'Eleos drop-in',
     'Climate change app',
   ]
-
-  const [calendarPickerWidth, setCalendarPickerWidth] = useState(0)
-
-  const handleSearch = (input: React.SetStateAction<string>) => {
-    console.log(input)
-  }
-
-  const handleSubmit = () => {
-    console.log('Submit')
-  }
 
   return (
     <SafeAreaView>
@@ -100,34 +131,40 @@ const EventSearchContainer = () => {
         <TopOfApp />
 
         <FreeSearchBar
-          handleSearch={handleSearch}
-          handleSubmit={handleSubmit}
+          handleChangeText={handleFreeTextChange}
+          handleSubmit={handleFreeTextSubmit}
         />
 
         <SectionView>
-          <EventSearchUpcomingQuickSearch />
+          <EventSearchQuickSearchUpcoming />
         </SectionView>
 
         <SectionView>
           <EventSearchCalendarPicker width={calendarPickerWidth} />
         </SectionView>
 
-        <Heading>Popular</Heading>
         <SectionView>
-          {PopularSearches.map((event, index) => (
-            <QuickSearchButton onPress={underDevelopmentAlert} key={index}>
-              <EventTitle>{event}</EventTitle>
-            </QuickSearchButton>
-          ))}
+          <EventSearchQuickSearch
+            choices={eventsSeriesChoices}
+            field="series"
+            heading="Series"
+          />
         </SectionView>
 
-        <Heading>Project Specific</Heading>
         <SectionView>
-          {ProjectSearches.map((event, index) => (
-            <QuickSearchButton onPress={underDevelopmentAlert} key={index}>
-              <EventTitle>{event}</EventTitle>
-            </QuickSearchButton>
-          ))}
+          {/*
+            TODO:
+            - Database: add project_res_id field to event
+            - API: add project_res_id to all events data
+            - API: update tests to include project_res_id
+            - App: get project specific quick search working
+            - App: get free text search working
+          */}
+          <EventSearchQuickSearch
+            choices={eventsSeriesChoices}
+            field="series"
+            heading="Project Specific"
+          />
         </SectionView>
       </ScrollView>
     </SafeAreaView>
