+++ app/src/Containers/index.ts
@@ -7,5 +7,5 @@ export { default as EventDetailContainer } from './EventDetailContainer'
 export { default as ProjectDetailContainer } from './ProjectDetailContainer'
 export { default as ProjectSearchResultsContainer } from './ProjectSearchResultsContainer'
 export { default as ProjectRegisterInterestContainer } from './ProjectRegisterInterestContainer'
-export { default as SearchContainer } from './SearchContainer'
-export {default as WelcomeContainer } from './WelcomeContainer'
+export { default as ProjectSearchContainer } from './ProjectSearchContainer'
+export { default as WelcomeContainer } from './WelcomeContainer'
